refactor: domain에 있던 datasource interface 이동
