+++ data/src/main/kotlin/team/duckie/app/android/data/_local/LocalModule.kt
@@ -23,7 +23,6 @@ import javax.inject.Singleton
 @InstallIn(SingletonComponent::class)
 @Module
 class DatabaseModule {
-
     @Singleton
     @Provides
     fun provideAppDatabase(@ApplicationContext context: Context): DuckieDatabase {
