+++ data/src/main/kotlin/team/duckie/app/android/data/auth/datasource/AuthDataSource.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.domain.auth.datasource
+package team.duckie.app.android.data.auth.datasource
 
 import team.duckie.app.android.domain.auth.model.AccessTokenCheckResponse
 import team.duckie.app.android.domain.auth.model.JoinResponse
