+++ data/src/main/kotlin/team/duckie/app/android/data/user/datasource/UserDataSource.kt
@@ -5,7 +5,7 @@
  * Please see full license: https://github.com/duckie-team/duckie-android/blob/develop/LICENSE
  */
 
-package team.duckie.app.android.domain.user.datasource
+package team.duckie.app.android.data.user.datasource
 
 import team.duckie.app.android.domain.category.model.Category
 import team.duckie.app.android.domain.tag.model.Tag
